+++ annotations/src/main/java/arez/annotations/Track.java
@@ -41,10 +41,12 @@ public @interface Track
 
   /**
    * Does the tracker mutate state or not.
+   * The default value us false thus making the transaction mode read-only. The intention is that tracker observers
+   * should primarily be reflecting Arez state to external systems. (i.e. views, network layers etc.)
    *
    * @return true if the tracker should run in a READ_WRITE transaction, false if it should run in a READ_ONLY transaction.
    */
-  boolean mutation() default true;
+  boolean mutation() default false;
 
   /**
    * Return true if the parameters should be reported to the core Arez runtime.